feature: User receiving result from an sse
In this commit i added an feature where the user will subscribe to a
server sent event using the endpoint /subscribe.

The user send data using a normal http request the server will add the
user data to a list of subscribed users.

The user data will be calculated and then sent to him in a form of a
server sent event.